dry-up did doc type and validation
