+++ packages/common-web/src/index.ts
@@ -9,3 +9,4 @@ export * from './ipld'
 export * from './types'
 export * from './times'
 export * from './strings'
+export * from './did-doc'
